# api/openapi.yaml
openapi: 3.1.0
info:
  title: AUSTRO OS API
  version: v1.0.0
  description: |
    AUSTRO OS - AI Content Operating System
    Every endpoint references a constitutional principle for enforcement.
servers:
  - url: http://localhost:8080
    description: Development server
paths:
...
  /api/auth/bootstrap:
    post:
      summary: Initialize the founder identity
      description: |
        Creates the single founder identity from the deployment configuration
        (AUSTRO_FOUNDER_USERNAME / AUSTRO_FOUNDER_PASSWORD). Succeeds exactly
        once; returns 409 once a founder exists or when bootstrap is not
        configured. There is no arbitrary self-registration.
      operationId: authBootstrap
      responses:
        '201':
          description: Founder initialized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: ["initialized"]
        '409':
          description: Already initialized or not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      principles:
        - Principle 9 — Security by Design
        - Principle 12 — Human Oversight
  /api/auth/login:
    post:
      summary: Authenticate with username and password
      description: |
        Issues an access token (15 minutes) and a refresh token (30 days,
        rotation enabled). Failures are uniform: they never reveal whether the
        username or the password was wrong.
      operationId: authLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      principles:
        - Principle 9 — Security by Design
        - Principle 10 — Privacy by Design
  /api/auth/refresh:
    post:
      summary: Rotate a refresh token
      description: |
        Validates the presented refresh token, rejects reused or revoked
        tokens, rotates it, and re-issues an access token for the same identity.
      operationId: authRefresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Rotated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid, reused or revoked token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      principles:
        - Principle 9 — Security by Design
  /api/auth/logout:
    post:
      summary: Revoke a refresh token
      description: Terminates the session held by the presented refresh token. Idempotent.
      operationId: authLogout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '204':
          description: Session terminated
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      principles:
        - Principle 9 — Security by Design
        - Principle 10 — Privacy by Design
  /api/me:
    get:
      summary: Current user profile
      description: |
        Returns the authenticated caller's safe profile, including the identity's
        authorization role (founder | workspace_admin | workspace_member). The
        role is read from the persisted identity at token-issuance time; it is
        never accepted from the client.
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      principles:
        - Principle 9 — Security by Design
        - Principle 10 — Privacy by Design
  /workspaces:
    get:
      summary: List workspaces
      description: |
        Organization-level workspace listing. Explicitly founder-only: a founder
        may enumerate every workspace in the organization; a workspace_admin may
        not (it reaches only its own workspace through GET /workspaces/{id})
        and a workspace_member is denied. There is no client-supplied workspace
        context for this operation.
      operationId: listWorkspaces
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workspace'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      principles:
        - Principle 1 — Vision First
        - Principle 9 — Security by Design
    post:
      summary: Create workspace
      description: |
        Create a new workspace. Explicitly founder-only and organization-level:
        only the founder carries the permission, so workspace_admin and
        workspace_member identities are denied. The name is required, trimmed
        and bounded to 100 characters; unknown request fields (including any
        client-supplied workspace_id) are rejected. A duplicate name returns a
        deterministic 409 conflict.
      operationId: createWorkspace
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: Workspace name
      responses:
        '201':
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Workspace name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      principles:
        - Principle 1 — Vision First
        - Principle 5 — Modular Design
  /workspaces/{id}:
    get:
      summary: Get workspace
      description: |
        Read a workspace by id. A workspace_admin may read only its own
        workspace: the workspace in the path must equal the admin's verified
        token workspace and never comes from a client-controlled source. A
        workspace_member is denied. A founder's organization-level read is
        explicit.
      operationId: getWorkspace
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Workspace details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          description: Invalid workspace id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      principles:
        - Principle 9 — Security by Design
        - Principle 10 — Privacy by Design
...
  /audit/verification:
    get:
      summary: Verify the audit hash chain
      description: |
        Re-reads the persisted chain and recomputes every link, reporting
        whether it is intact and how many events were checked. The check is
        independent of anything the writer remembers, so it detects an edit made
        behind the application's back.

        Founder only, and expensive by nature: it reads the whole table, because
        a bounded proof of a hash chain is not a thing. It is meant to be run
        deliberately, not polled.
      operationId: verifyAuditChain
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Result of the whole-chain integrity check
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditVerification'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Caller is not the founder
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      principles:
        - Principle 11 — Human Oversight
        - Principle 9 — Security by Design
...
  /publications/{id}:
    get:
      summary: Get one publication status
      operationId: getPublication
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PublicationID'
      responses:
        '200':
          description: Publication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Publication'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      principles:
        - Principle 10 — Privacy by Design
...
components:
  parameters:
    PublicationID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    Workspace:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: integer
          description: Error code
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: Bearer access token (15 minute lifetime)
        token_type:
          type: string
          enum: ["bearer"]
        expires_in:
          type: integer
          description: Access token lifetime in seconds
        refresh_token:
          type: string
          description: Rotation-enabled refresh token
    MeResponse:
      type: object
      description: |
        The authenticated caller's safe profile. Only profile fields and the
        identity's authorization role are returned; password hashes,
        refresh-token material and internal database fields are never exposed.
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        display_name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: ["founder", "workspace_admin", "workspace_member"]
        is_founder:
          type: boolean
        workspace_id:
          type: string
          format: uuid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...